# optimize memory usage in meantimer