--- a/RecoLocalMuon/DTSegment/src/DTSegmentCand.h
+++ b/RecoLocalMuon/DTSegment/src/DTSegmentCand.h
@@ -96,6 +96,7 @@ class DTSegmentCand{
     virtual void setDirection(LocalVector& dir) { theDirection = dir; }
 
     /// add hits to the hit list.
+    virtual void add(AssPoint newHit) ;
     virtual void add(std::shared_ptr<DTHitPairForFit> hit, DTEnums::DTCellSide code) ;
 
     /// remove hit from the candidate
@@ -119,9 +120,9 @@ class DTSegmentCand{
 
     /// number of different layers with hits
     virtual int nLayers() const ;
-    
+
     /// the used hits
-    virtual AssPointCont hits() const { return theHits;}
+    virtual const AssPointCont& hits() const { return theHits;}
 
     /// convert this DTSegmentCand into a DTRecSegment2D
     //  DTSLRecSegment2D* convert() const;