addressing the issue reported in https://hypernews.cern.ch/HyperNews/CMS/get/recoDevelopment/1314.html
optimize memory usage in DT local reco (algorithm output unchanged)