@cmsbuild 2015-02-03
A new Pull Request was created by @ptraczyk (Piotr Traczyk) for CMSSW_7_4_X.

optimize memory usage in meantimer

It involves the following packages:

RecoLocalMuon/DTSegment


@cmsbuild, @cvuosalo, @nclopezo, @slava77 can you please review it and eventually sign? Thanks.
@bellan this is something you requested to watch as well.
You can sign-off by replying to this message having '+1' in the first line of your reply.
You can reject by replying  to this message having '-1' in the first line of your reply.
If you are a L2 or a release manager you can ask for tests by saying 'please test' in the first line of a comment.