@slava77 2015-02-03
@cmsbuild please test

Piotr, thanks for wrapping up with this.