@cmsbuild 2015-02-03
The tests are being triggered in jenkins.