@cmsbuild 2015-02-03
+1
Tested at: 061e64082ea25a76f18875a486d4f3a8eadddf0d
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-7522/2325/summary.html