@cmsbuild 2015-02-03
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-7522/2325/summary.html