@cvuosalo 2015-02-04
Started extended test for muon workflow.