@cvuosalo 2015-02-06
+1

For #7522 061e64082ea25a76f18875a486d4f3a8eadddf0d

Small code change to reduce unnecessary copying of data values and thereby reduce CPU time.

Code changes are satisfactory.

Jenkins tests show no problems or significant differences.

Workflow 20.0 (muons) was run as an extended test (2000 events) to check timing differences. The new code shows significant improvement:

  delta/mean   | delta/orJob     |   original         |    new   |  module name
  ----------      |    ------------        |     --------       |          ----    |   ------------
   -0.500873     | -0.69%   |     16.47 ms/ev     |    9.87 ms/ev | dt4DSegments
   -0.498409   |   -0.59%       | 14.13 ms/ev     |     8.49 ms/ev | dt4DCosmicSegments

Profiling also shows about a 1/3 reduction in time used by the DTMeantimer (going from a reported 2.3% of total time to 1.5%).